+++ plugins/O-Formant/Source/PluginEditor.h
@@ -43,6 +43,8 @@ private:
     std::unique_ptr<juce::WebSliderRelay> pitchGlideRelay;
     std::unique_ptr<juce::WebSliderRelay> transitionTimeRelay;
     std::unique_ptr<juce::WebSliderRelay> singersFormantRelay;
+    std::unique_ptr<juce::WebSliderRelay> nasalCouplingRelay;
+    std::unique_ptr<juce::WebSliderRelay> nasalPlaceRelay;
     std::unique_ptr<juce::WebSliderRelay> outputGainRelay;
     std::unique_ptr<juce::WebSliderRelay> stereoWidthRelay;
 
@@ -76,6 +78,8 @@ private:
     std::unique_ptr<juce::WebSliderParameterAttachment> pitchGlideAttachment;
     std::unique_ptr<juce::WebSliderParameterAttachment> transitionTimeAttachment;
     std::unique_ptr<juce::WebSliderParameterAttachment> singersFormantAttachment;
+    std::unique_ptr<juce::WebSliderParameterAttachment> nasalCouplingAttachment;
+    std::unique_ptr<juce::WebSliderParameterAttachment> nasalPlaceAttachment;
     std::unique_ptr<juce::WebSliderParameterAttachment> outputGainAttachment;
     std::unique_ptr<juce::WebSliderParameterAttachment> stereoWidthAttachment;